Cleaned up FreeformCompare
@tanczosm

An MLDataSet needs to be initialized with some sort of constructor, so I
declared two new arrays when initializing the trainingSet.
Trainers are declared and initialized on the same line for readability.

We should look into the trainToError utility rather than have a set
number of iterations.

I'll look into whether or not we can overwrite data and keep the
existing training weights, just wanted to share what I had.