FreeformCompare adds new data while training
It's a bit sloppy, but you can see that at iteration 100 it trains
accurately with the first two data sets. At iteration 101 new data is
introduced and at iteration 300, it trains accurately with the third set
in the mix. The only problem is that the basic network trains a lot
faster than the freeform network, so we can't exactly check the accuracy
by comparing the two.